Rename auth-annotation to radar-auth
- New library will still have an annotation for authenticating the token and checking it's scope
- New library will have a utility class for checking permissions
