+++ radar-auth/bin/radar-is.yml
@@ -1,0 +1,4 @@
+username: test_username
+password: test_password
+tokenValidationEndpoint: http://localhost:8089/oauth/check_token
+publicKeyEndpoint: http://localhost:8089/oauth/token_key
