+++ radar-auth/src/main/java/org/radarcns/auth/exception/ConfigurationException.java
@@ -1,0 +1,18 @@
+package org.radarcns.auth.exception;
+
+/**
+ * Created by dverbeec on 20/09/2017.
+ */
+public class ConfigurationException extends RuntimeException {
+    public ConfigurationException() {
+        super();
+    }
+
+    public ConfigurationException(String message) {
+        super(message);
+    }
+
+    public ConfigurationException(Exception cause) {
+        super(cause);
+    }
+}
