+++ radar-auth/src/main/java/org/radarcns/auth/exception/TokenValidationException.java
@@ -1,0 +1,18 @@
+package org.radarcns.auth.exception;
+
+/**
+ * Created by dverbeec on 15/09/2017.
+ */
+public class TokenValidationException extends RuntimeException {
+    public TokenValidationException() {
+        super();
+    }
+
+    public TokenValidationException(String message) {
+        super(message);
+    }
+
+    public TokenValidationException(Exception cause) {
+        super(cause);
+    }
+}
