+++ radar-auth/src/test/resources/radar-is.yml
@@ -1,0 +1,2 @@
+publicKeyEndpoint: http://localhost:8089/oauth/token_key
+resourceName: unit_test
