+++ settings.gradle
@@ -1,3 +1,3 @@
 rootProject.name = 'management-portal'
-include 'auth-annotation'
 include 'oauth-client-util'
+include 'radar-auth'
